refactor(project-key): SSOT 化して projectKey 算出を共通 helper に集約
intent(project-key): 同じ `<repoName>-<sha256(realpath)[0..12]>` 計算が ProjectConfigStore / TaskStore / WorktreeOps の 3 箇所に重複していた。形式が乖離すると `~/.config/gozd/projects/<projectKey>/` と `~/.local/share/gozd/worktrees/<projectKey>/` が分裂するため、SSOT として helper に抽出する
decision(api-shape): 用途で 2 関数を提供する。`compute(forMainRepoRoot:)` は呼び出し側が main repo root を保証するケース用、`resolveAndCompute(for:)` は worktree 配下のどの dir からでも呼ばれるケース用（git rev-parse --git-common-dir で main repo root に解決してから compute）。各 store の既存挙動を維持する
constraint(format-stability): projectKey 形式を変えると全 store の保存先が変わるため、変更時は移行コードが必須。helper のコメントに明記
