+++ apps/native/Sources/GozdCore/ProjectConfigStore.swift
@@ -1,10 +1,9 @@
-import CryptoKit
 import Foundation
 import GozdProto
 import SwiftProtobuf
 
 // プロジェクト固有の設定永続化（`~/.config/gozd/projects/<projectKey>/config.json`）。
-// projectKey は dir realpath の SHA-256 先頭 12 文字 + repoName。
+// projectKey の算出は `ProjectKey` を参照。
 //
 // 将来バージョンが増やしたフィールドが入った JSON を旧 binary で読んでも
 // parse error にならないよう `ignoreUnknownFields = true` を渡す。
@@ -37,11 +36,7 @@ public final class ProjectConfigStore {
   }
 
   private func filePath(for dir: String) -> String {
-    let realpath = (dir as NSString).resolvingSymlinksInPath
-    let repoName = (realpath as NSString).lastPathComponent
-    let digest = SHA256.hash(data: Data(realpath.utf8))
-    let hash = digest.compactMap { String(format: "%02x", $0) }.joined()
-    let projectKey = "\(repoName)-\(String(hash.prefix(12)))"
+    let projectKey = ProjectKey.compute(forMainRepoRoot: dir)
     return (configDir as NSString)
       .appendingPathComponent("projects")
       .appending("/\(projectKey)/config.json")
